Art For Test Level
That's right, the test level art is finished!  i have posted the
complete level design along with the components and an explanation as to
what you might get confused on.  if you need me to change anything, or
to explain something, e-mail me at nburns12@uncc.edu and i will get back
to you asap.