Player Weapons and Attacks
Includes a sword with a stab, slash, and air spin attack.  It also has a
bow that shoots simple arrows.  I made a simple enemy that can get hit
by the sword/arrows, take damage and get destroyed at 0 health.  Press 1
in-game for sword and 2 for the bow. A is to stab/shoot arrows, S is to
slash/air spin attack.