Basic AI
We need to do more work with collisions (the player object works just
fine, but Streeper was right...we need a sort of "physics" architecture
which can work with both player and NPCs).

This particular version of the project which I am submitting has an AI
which paces back and forth and will aggressively follow the player if he
comes to close.

Tomorrow I will work more on the physics and try to integrate
Carpenter's weapons system into the enemy NPCs. We definitely need to
meet this Thursday and Sunday.