Basic Physics
Moving Platforms (you can stand on these!)
Objects that fall realistically

Notice: I had to modify some code for the arrow/Player so that it would
interact with the obj_platform the same way. ("par_platform" to
"obj_Static_Parent" in the place_meeting functions)

Todo
Need to consider if we should bump up the fps of the room to 60 - would
make collisions/physics better.
Need to work on collision with the dynamic physic objects.
Need to make moving platforms interact with any object, not just marker
instances.
Player does not move with the moving platforms though he should.