Slightly organized files and started on rope swing
I put all the weapon objects/sprites in their own folders and created a
rope swing object  from the 3 rope swing parts and made the collision
only on the blade.

Note:  If we decide to switch to 60fps, we need to decide soon as it
will alter how fast the player/platforms move as well as weapon attacks
and possible objects we haven't added in yet.