updated physics player
!. I deleted my folder i saw no sense in us have 5 folders of the full
game.  Because i am currently only working on collision I will only be
uploading the player.  Note the player object may be converted to a
parent class so further objects can inherit these physics. This
separation will be done at a later time.
2.  Because i am only working on one thing i will only be uploading the
object i am working on as to cut down on the memory size of our repo.